fixed: move util function to utils
