+++ src/components/Modal.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import styled from "styled-components";
 
-import { useWindowDimensions } from "../global/util"
+import { useWindowDimensions } from "../global/utils"
 import { COLOR } from "../global/const"
 
 export const Modal = ({
