+++ src/containers/PracticeAnswer/PracticeAnswer.js
@@ -27,7 +27,7 @@ import { backward, forward } from "./PracticeAnswerHelper";
 // Global
 import { Body, Header } from "../../components/Typography";
 import { COLOR } from "../../global/const";
-import { useWindowDimensions } from "../../global/util";
+import { useWindowDimensions } from "../../global/utils";
 
 const CONTAINER_PADDING = 64;
 const NAVBAR_HEIGHT = 54;
